requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
Django==3.0
djangorestframework==3.11.2
idna==2.10
pkg-resources==0.0.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
